+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tile decoding is now implemented.
 - Speed up `Bitmap::pixels()` for large symbols by about factor six. The performance
   should now be linear in the number of pixels.
+- A `DataMatrixBuilder` is now available that allows to specify allowed encodation types.
 
 ### Fixes